admin verification to backend requests

+++ backend/routes/admin.ts
@@ -21,7 +21,16 @@ router.get('/users', async (req: Request, res: Response) => {
 });
 
 router.delete('/users/:id', async (req: Request, res: Response) => {
-  const { id } = req.params;
+  const id = parseInt(String(req.params.id));
+
+  if (isNaN(id)) {
+    return res.status(400).json({ error: 'Invalid User Id'})
+  }
+
+  if (id === req.user!.user_id) {
+    return res.status(400).json({ error: 'You cannot delete your own account'});
+  }
+
   try {
     const result = await pool.query(
       `Delete FROM users
@@ -30,20 +39,30 @@ router.delete('/users/:id', async (req: Request, res: Response) => {
       [id]
     );
     if (result.rows.length === 0) {
-      return res.status(404).json({error: 'User not found' });
+      return res.status(404).json({ error: 'User not found' });
     }
-    return res.status(200).json({message: 'User deleted', deleted: result.rows[0] })
+    return res.status(200).json({ message: 'User deleted', deleted: result.rows[0] })
   } catch (err) {
     console.error('Error deleting users:', err);
-    return res.status(500).json({error: 'Failed to delete user' });
+    return res.status(500).json({ error: 'Failed to delete user' });
   }
 });
 
 router.patch('/users/:id/role', async (req: Request, res: Response) => {
-  const { id } = req.params;
+  const id = parseInt(String(req.params.id));
   const { role } = req.body;
   if (!['admin', 'user'].includes(role)) {
-    return res.status(400).json({error: 'Role must be admin or user'});
+    return res.status(400).json({ error: 'Role must be admin or user' });
+  }
+
+  if (isNaN(id)) {
+    return res.status(400).json({ error: 'Invalid User Id' });
+  }
+  if (id === req.user!.user_id) {
+    return res.status(400).json({ error: 'You cannot change your own role' })
+  }
+  if (!role || !['admin','user'].includes(role)) {
+    return res.status(400).json({error: 'Role must be admin'})
   }
 
   try {
